Fix flaw escaping multiple embed delims
`escapeEmbedContent` was erroneously adding a new slash before the
last character of the `needsEscapesPattern`, but since (as documented
on `needsEscapesPattern`!) that pattern matches trailing slashes, this
is flawed.

Rather than before the last character of the match, we need to add the
escaping slash after the _first_ character of the match.  That character
is guaranteed to be the first char of the embed-end delim we're trying
to escape.