Ensure the embedContentNode StringNode is correct
`QuotedStringNode` demands that the `String`s it is initialized with
be a JSON-escaped, and it documents this expectation (this demand helps
avoid unnecessary escaping/unescaping on strings), but everywhere we
handled these as part of `EmbedBlockNode.embedContentNode`, we were
carefully ensuring the opposite.

Since we ensured the opposite consistently on read _and_ write, we
currently have two wrongs making a right.  This commits flips things
around so that the QuotedStringNode interface demands are being
respected to avoid confusion in the future.

This commit also notes that the handling of `embedTag` and `metadataTag`
is also strictly-speaking incorrect.  These will be looked at in the
next commit.