Embed preamble JSON interoperability
It's very important that we be JSON equivalent (modulo comments), and
one spot this was leaky was in our embed tags and embed metadata. We
need them to be equivalent to JSON strings so they do their part in
making embed blocks fully equivalent to:

```
{
  "embedTag": "the tag content",
  "embedMetadata": "the metadata content",
  "embedContent": "EMBED CONTENT!"
}
```

We now properly parse embed tags and embed metadata (nearly) identically
to KSON/JSON strings, ensuring that this equivalence holds.

The "nearly" noted in the previous sentence alludes to the fact that we
allow these to be syntactically quote-free, with their "end quotes"
being the `:`that ends an embed tag or the newline that ends the whole
embed preamble. Though this feels fairly novel, the way the code clicks
together in a reassuringly clean and clear way in this change is a very
strong indicator we get to have our syntax cake and eat it too.